config: use per-post title permalink and keep pretty in _config
The title permalink in _config breaks the tags even though they work
locally. No idea why that happens, but this solution works as expected
and given that I always copy the header from an existing post, it is not
tedious.

Maybe one day I will find out why this happens... or not.